gcc-lispc: lambdas!
Support lambdas, and through them, closures!

This requires updating several components.

The program 'regions' are updated from two hard-coded regions, to
generated region indices.  Each function is given a region into which it
emits code, preventing instructions from lambdas from being mixed in
with the surrounding function.

Items in the environment frame are simpler reserved as indices.  When
referenced, the depth and index of the reference are used to generate an
LD instruction.  Undefined references produce much more readable error
messages.

The 'funcall' "macro" is used to call closures, as they cannot be
directly called.