lambda-ai: IT RUNS!
The searching LambdaMan AI now runs and will successfully wander around
the map, though it does like the upper right corner.

There are several components to this patch:

0. Several of the tree helper functions have atom checks, for sanity
checking the input.

1. The root cause of the initial crash was the CAR in best-move.
best-move-yet returns an int, not a cons.

2. +y is down, not up, so next-y had to be swapped

3. enumerate-moves was missing bounds checks, so it would enumerate
moves off the edge of the map, causing a crash.  Add bounds checking to
next-x and next-y, which will return the same x and y if the next value
would go off the map.